ecommerce/requirements.txt
asgiref==3.8.1
attrs==21.2.0
Automat==20.2.0

bcrypt==3.2.0
blinker==1.4
certifi==2024.7.4
chardet==4.0.0
click==8.1.3
cloud-init==25.1.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.9
constantly==15.1.0
cryptography==44.0.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-anymail==10.3
sentry-sdk==2.32.0
# Optional for AWS CloudWatch logging in production:
# boto3==1.34.90
# watchtower==3.0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1

jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.2
pillow==11.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.1


PyHamcrest==2.0.2
PyJWT==2.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
pytz==2022.1

requests==2.31.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sqlparse==0.5.3
svgwrite==1.4.3

Tree==0.2.4

typing_extensions==4.14.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.5.0
wadllib==1.3.6
whitenoise==6.9.0
zipp==3.19.1

setuptools>=78.1.1 # pinned by Snyk to avoid a vulnerability

# --- DEV/LOCAL ONLY ---
# pytest==7.4.4
# coverage==7.4.4
# black==24.3.0
# isort==5.12.0
# Add more dev/test tools above as needed (not installed in production)
celery==5.4.0
redis==5.0.1